+++ bruges/models/wedge.py
@@ -2,14 +2,15 @@ from collections import namedtuple
 
 import numpy as np
 import scipy.ndimage as sn
+import scipy.ndimage.morphology as morph
 
-from bruges.util import sigmoid
+from bruges.util import sigmoid, root, power
 
 
 def pad_func(before, after):
     """
-    Padding function. Operates on vector *in place*,
-    as per the np.pad documentation.
+    Padding function. Operates on vector *in place*, per the np.pad
+    documentation.
     """
     def pad_with(x, pad_width, iaxis, kwargs):
         x[:pad_width[0]] = before[-pad_width[0]:]
@@ -18,23 +19,27 @@ def pad_func(before, after):
     return pad_with
 
 
-def get_strat(strat, thickness, mode='stretch', kind='nearest', position=1, wedge=None, zoom_mode='nearest'):
+def get_strat(strat,
+              thickness,
+              kind='nearest',
+              position=1,
+              wedge=None,
+              zoom_mode='nearest'
+              ):
     """
     Take a 'stratigraphy' (either an int, a tuple of ints, or a list-like of
     floats) and expand or compress it to the required thickness.
-    
-    'mode' can be 'stretch' or 'crop'
-    
+
     `kind` can be 'nearest', 'linear', 'quadratic', or 'cubic'.
     """
     orders = {'nearest': 0, 'linear': 1, 'quadratic': 2, 'cubic': 3}
     order = orders.get(kind, 0)
-    
-    if isinstance(strat, int) and order==0:
+
+    if isinstance(strat, int) and (order == 0):
         out = np.repeat([strat], thickness)
-    elif isinstance(strat, float) and order==0:
+    elif isinstance(strat, float) and (order == 0):
         out = np.repeat([strat], thickness)
-    elif isinstance(strat, tuple) and order==0:
+    elif isinstance(strat, tuple) and (order == 0):
         out = np.repeat(strat, int(round(thickness/len(strat))))
     else:
         if position == 0:
@@ -45,7 +50,7 @@ def get_strat(strat, thickness, mode='stretch', kind='nearest', position=1, wedg
             strat = strat[:int(thickness/wedge_zoom)]
         zoom = thickness / len(strat)
         out = sn.zoom(strat, zoom=zoom, order=order, mode=zoom_mode)
-       
+
     # Guarantee correct length by adjusting bottom layer.
     missing = int(np.ceil(thickness - out.size))
     if out.size > 0 and missing > 0:
@@ -66,7 +71,10 @@ def get_conforming(strat, thickness, conformance):
     if strat.size == thickness:
         return strat
     elif strat.size > thickness:
-        return strat[:thickness] if conformance == 'top' else strat[-thickness:]
+        if conformance == 'top':
+            return strat[:thickness]
+        else:
+            return strat[-thickness:]
     else:
         if conformance == 'top':
             return np.pad(strat, [0, thickness-strat.size], mode='wrap')
@@ -75,62 +83,131 @@ def get_conforming(strat, thickness, conformance):
     return
 
 
+def get_subwedges(target, breadth):
+    """
+    For a binary target (the reference trace of a wedge),
+    create the range of net:gross subwedges. We do this
+    with binary morphologies in the following way:
+
+    - Erode the 'secondary' component (whatever appears
+      second in the target array) one step at a time,
+      until there is nothing left and the resulting
+      trace contains only the primary.
+    - Dilate the secondary component until there is
+      nothing left of the primary and the resulting trace
+      contains only the secondary.
+    - Arrange the traces in order, starting with all
+      primary, and ending in all secondary. The target
+      trace will be somewhere in between, but not
+      necessarily in the middle.
+
+    Returns a 2D array with one target wedge trace per
+    section.
+
+    Args:
+        target (array): A 1D array length N, the 'reference'
+            trace for the wedge. The reference trace has
+            thickness '1' in the wedge model. This trace must
+            be 'binary' — i.e. it must contain exactly 2
+            unique values.
+        breadth (int): How many new reference traces should
+            be in the output.
+
+    Returns:
+        tuple (ndarray, ndarray, int). The ndarray has shape
+            N x breadth. It represents one target wedge trace
+            per section in 'breadth'. The integer is the
+            position of the target trace in the ndarray's
+            second dimension.
+    """
+    try:
+        components = a, b = np.unique(target)
+    except ValueError:
+        raise ValueError("Must be a binary (2-component) wedge.")
+
+    out = [target]
+
+    temp = target.copy()
+    while b in temp:
+        ero = morph.binary_erosion(temp == b)
+        temp = components[ero.astype(int)]
+        out.append(temp)
+
+    out = out[::-1]
+    ref = len(out)
+
+    temp = target.copy()
+    while a in temp:
+        dil = morph.binary_dilation(temp == b)
+        temp = components[dil.astype(int)]
+        out.append(temp)
+
+    arr_ = np.array(out).T
+    h, w = arr_.shape
+    arr = sn.zoom(arr_, zoom=(1, breadth/w), mode='nearest')
+
+    ng = np.divide(np.sum(arr == a, axis=0), target.size)
+
+    return arr, ng, ref * breadth / w
+
+
 def wedge(depth=(30, 40, 30),
           width=(10, 80, 10),
-          breadth=None,  # Not implemented.
+          breadth=None,
           strat=(0, 1, 2),
           thickness=(0.0, 1.0),
           mode='linear',
-          conformance='proportional',
-         ):
+          conformance='both',
+          ):
     """
     Generate a wedge model.
-    
+
     Args:
-        depth (int or tuple): The vertical size of the model. If a 3-tuple, then
-            each element corresponds to a layer. If an integer, then each layer
-            of the model will be 1/3 of the thickness. Note that if the 'right'
-            wedge thickness is more than 1, then the total thickness will be
-            greater than this value.
+        depth (int or tuple): The vertical size of the model. If a 3-tuple,
+            then each element corresponds to a layer. If an integer, then each
+            layer of the model will be 1/3 of the thickness. Note that if the
+            'right' wedge thickness is more than 1, then the total thickness
+            will be greater than this value.
         width (int or tuple): The width of the model. If a 3-tuple, then each
-            element corresponds to a 'zone' (left, middle, right). If an integer,
-            then the zones will be 10%, 80% and 10% of the width, respectively.
-            Note: for small models, you'll get better results with an odd number
-            of pixels in the middle (wedge) zone.
-        breadth (int or tuple): Not implemented. Raises an error.
+            element corresponds to a 'zone' (left, middle, right). If an
+            integer, then the zones will be 10%, 80% and 10% of the width,
+            respectively.
+        breadth (None or int): Not implemented. Raises an error.
         strat (tuple): Stratigraphy above, in, and below the wedge. This is the
-            'reference' stratigraphy. If you give integers, you get 'solid' layers
-            containing those numbers. If you give arrays, you will get layers of
-            those arrays, expanded or squeezed into the layer thicknesses implied
-            in `depth`. 
+            'reference' stratigraphy. If you give integers, you get 'solid'
+            layers containing those numbers. If you give arrays, you will get
+            layers of those arrays, expanded or squeezed into the layer
+            thicknesses implied in `depth`.
         thickness (tuple): The wedge thickness on the left and on the right.
             Default is (0.0, 1.0) so the wedge will be thickness 0 on the left
-            and the wedge thickness specified in the depth argument on the right.
-            If the thickness are equal, you'll have a flat, layer-cake model.
-        mode (str): What kind of interpolation to use. Default: 'linear'. Other
-            option is 'clinoform' (or synonymously 'sigmoid'), which makes a
-            clinoform-like body.
-        conformance (str): 'top', 'bottom', or 'proportional' (the default,
-            although you can use any word other than 'top', 'bottom' or 'base').
-            How you want the layers inside the wedge to behave. For top and
-            bottom conformance, if the layer needs to be thicker than the
-            reference position it will be wrapped (i.e. repeated) using the
-            `'wrap'` mode of `np.pad()`.
+            and the wedge thickness specified in the depth argument on the
+            right. If the thickness are equal, you'll have a flat, layer-cake
+            model.
+        mode (str or function): What kind of interpolation to use. Default:
+            'linear'. Other options are 'sigmoid', which makes a clinoform-like
+            body, 'root', which makes a steep-sided bowl shape like the edge of
+            a channel, and 'power', which makes a less steep-sided bowl shape.
+            If you pass a function, give it an API like np.linspace: f(start,
+            stop, num), where start is the left thickness, stop is the right
+            thickness, and num is the width of the middle (wedge) 'zone'.
+        conformance (str): 'top', 'bottom', or 'both' (the default). How you
+            want the layers inside the wedge to behave. For top and bottom
+            conformance, if the layer needs to be thicker than the reference.
 
     Returns:
         namedtuple[ndarray, ndarray, ndarray, int]: A tuple containing the
             2D wedge model, the top 'horizon', the base 'horizon', and the
             position at which the wedge has thickness 1 (i.e. is the thickness
             specfied by the middle layer depth and/or strat).
-
-    TODO
-    - Nearest interp ints, but linear interp floats (e.g. rock properties).
-    - Breadth argument implements the third dimension.
-    - If the wedge layer is a tuple of two ints, e.g. (1, 2, 1, 2, 1), then
-        you are making a 'binary wedge', which has special features.
     """
-    if breadth is not None:
-        raise NotImplementedError("The breadth argument is not implemented yet.")
+    # Decide if binary (2 rocks in the wedge).
+    if np.unique(strat[1]).size == 2:
+        binary = True
+    else:
+        binary = False
+
+    if breadth is None or breadth == 0:
+        breadth = 1
 
     # Allow wedge to be thin-thick or thick-thin.
     left, right = thickness
@@ -140,6 +217,7 @@ def wedge(depth=(30, 40, 30),
     else:
         flip = False
 
+    # Get all layers thicknesses.
     if isinstance(depth, int):
         L1, L2, L3 = 3 * [depth//3]  # Sizes if depth is just a number.
         L3 += 1
@@ -147,39 +225,61 @@ def wedge(depth=(30, 40, 30),
         L1, L2, L3 = map(int, depth)
     L3 += int(right * L2)  # Adjust bottom layer.
 
+    # Get all zone widths.
     if isinstance(width, int):
-        Z1, Z2, Z3 = width // 10, int(0.8 * zones), zones // 10
+        Z2 = round(0.8 * width)
+        Z1 = round(max(1, width/10))
+        Z3 = width - Z2 - Z1
+        width = int(Z1), int(Z2), int(Z3)
     else:
         Z1, Z2, Z3 = width  # Z1 and Z3 are the bookends.
 
-    if mode == 'linear':
-        zooms = np.linspace(left, right, Z2)
-    elif mode in ['clinoform', 'sigmoid']:
-        zooms = sigmoid(left, right, 1+Z2)
-    else:
-        raise TypeError("Mode not recognized.")
+    # Deal with different interpolation patterns.
+    modes = {
+        'linear': np.linspace,
+        'clinoform': sigmoid,
+        'sigmoid': sigmoid,
+        'root': root,
+        'power': power,
+    }
+    zooms = modes.get(mode, mode)(left, right, Z2)
 
     # Get the reference stratigraphy in each layer.
     # The 'well log' case is tricky, because layer1 and layer3
     # need to know about the amount of zoom on the wedge layer.
     # There must be an easier way to do this.
     layer1 = get_strat(strat[0], L1, position=0, wedge=(strat[1], L2))
-    layer2 = get_strat(strat[1], L2, position=1)
+    layer2_ = get_strat(strat[1], L2, position=1)
     layer3 = get_strat(strat[2], L3, position=-1, wedge=(strat[1], L2))
 
+    # Deal with width. We discard the reference breadth.
+    if binary and (breadth >= 2):
+        layer2s, _n2g, _ = get_subwedges(layer2_, breadth=breadth)
+        layer2s = layer2s.T
+    else:
+        layer2s, _ = [layer2_], None
+
+    # Make the padding function.
     padder = pad_func(layer1, layer3)
 
-    # Collect wedge pieces, then pad top & bottom, then stack, then pad left & right.
-    if conformance in ['top', 'bottom', 'base']:
-        wedges = [get_conforming(layer2, z*L2, conformance) for z in zooms]
-    else:
-        wedges = [get_strat(layer2, thickness=z*L2) for z in zooms]
-    padded = [np.pad(w, [L1, L3-w.size], mode=padder) for w in wedges] 
-    wedge = np.pad(np.stack(padded), [[Z1, Z3], [0, 0]], mode='edge')
+    # For everything in breadth:
+    model = []
+    for layer2 in layer2s:
+        # Collect wedge pieces, then pad top & bottom, then stack,
+        # then pad left & right.
+        if conformance in ['top', 'bottom', 'base']:
+            wedges = [get_conforming(layer2, z*L2, conformance) for z in zooms]
+        else:
+            wedges = [get_strat(layer2, thickness=z*L2) for z in zooms]
+        padded = [np.pad(w, [L1, L3-w.size], mode=padder) for w in wedges]
+        wedge = np.pad(np.stack(padded), [[Z1, Z3], [0, 0]], mode='edge')
+        model.append(wedge.T)
+    model = np.array(model)
 
     # Make the top and base 'horizons'.
-    top = np.ones(np.sum(width)) * L1
-    base = np.pad(L1 + zooms * L2, [Z1, Z3], mode='edge')
+    top = np.repeat((np.ones(np.sum(width)) * L1)[:, None], breadth, axis=-1)
+    base_ = np.pad(L1 + zooms * L2, [Z1, Z3], mode='edge')
+    base = np.repeat(base_[:, None], breadth, axis=-1)
 
     # Calculate the reference profile ('well' position).
     if left <= 1 <= right:
@@ -189,10 +289,20 @@ def wedge(depth=(30, 40, 30),
     else:
         ref = -1
 
+    # Flip if required.
     if flip:
-        wedge = np.flipud(wedge)
+        model = np.flip(model, axis=2)
         base = base[::-1]
         ref = sum(width) - ref
 
+    # Move the 'breadth' dim to last.
+    if model.shape[0] > 1:
+        model = np.moveaxis(model, 0, -1)
+
+    # Build and return output.
     Wedge = namedtuple('Wedge', ['wedge', 'top', 'base', 'reference'])
-    return Wedge(wedge.T, top, base, ref)
+    return Wedge(np.squeeze(model),
+                 np.squeeze(top),
+                 np.squeeze(base),
+                 ref
+                 )